Move demo screen, add labels, reorder
